Fixed controllers

+++ ApiShared/src/main/scala/uk/ac/ncl/openlab/intake24/api/shared/UserAdmin.scala
@@ -2,7 +2,7 @@ package uk.ac.ncl.openlab.intake24.api.shared
 
 import uk.ac.ncl.openlab.intake24.services.systemdb.admin.{SurveyUser, SurveyUserAlias, UserInfoWithId}
 
-case class DeleteUsersRequest(userIds: Seq[Int])
+case class DeleteUsersRequest(userIds: Seq[Long])
 
 case class DeleteSurveyUsersRequest(surveyId: String, userNames: Seq[String])
