+++ CommonSQL/src/main/scala/anorm/AnormUtil.scala
@@ -11,7 +11,7 @@ object AnormUtil {
   def isNull(row: Row, columnName: String): Boolean = row.get(columnName).get._1 == null
 
   @deprecated("Create a pull request for anorm instead...")
-  def batchKeys(sql: BatchSql)(implicit conn: Connection): Seq[Int] = {
+  def batchKeys(sql: BatchSql)(implicit conn: Connection): Seq[Long] = {
     
     val stmt = sql.getFilledStatement(conn, true)
     val result = stmt.executeBatch()
@@ -22,10 +22,10 @@ object AnormUtil {
       if (result.exists(_ != 1))
         throw new RuntimeException("Failed batch update")
 
-      val buf = Buffer[Int]()
+      val buf = Buffer[Long]()
 
       while (keys.next()) {
-        buf += keys.getInt("id")
+        buf += keys.getLong("id")
       }
 
       buf
